Switch to iterative KV parsing (#91)
This fixes existing issues with KV parsing that were not easily solved
with the regular expression approach (notably escaping and empty
values). This new version is more efficient (searching the string only
once), and allows for empty values (#74, #75).

This also adds a new function `joinKVStringForGCloud`, which tries to
pick the "best" delimiter based on the input. It's fairly expensive
(since each iteration requires searching the entire KV space), but it
fixes bugs where users want to include values that have a comma or other
delimiter that gcloud will use by default.

Fixes #74 
Fixes #75